aaaaaaa

+++ src/db.rs
@@ -120,7 +120,7 @@ impl MainDatabase {
 
   pub async fn get_user(mut db: Connection<Self>, user_id: u32) -> Option<User> {
     let result = sqlx::query("SELECT user_id, username, email, password_hash, created_on, last_activity, user_role, token FROM users WHERE user_id = $1")
-      .bind(user_id)
+      .bind(user_id as i32)
       .fetch_optional(&mut *db).await
       .unwrap();
     result.map(|row| User {
